+++ R/dopri.R
@@ -108,6 +108,19 @@
 ##'
 ##' @param verbose Be verbose, and print information about each step.
 ##'   This may be useful for learning about models that misbehave.
+##'   Valid values are \code{TRUE} (enable debugging) or \code{FALSE}
+##'   (disable debugging) or use one of \code{dopri:::VERBOSE_QUIET},
+##'   \code{dopri:::VERBOSE_STEP} or \code{VERBOSE:::VERBOSE_EVAL}.
+##'   If an R function is provided as the argument \code{callback}
+##'   then this function will also be called at each step or
+##'   evaluation (see below for details).
+##'
+##' @param callback Callback function that can be used to make verbose
+##'   output more useful.  This can be used to return more information
+##'   about the evaluation as it proceeds, generally as information
+##'   printed to the screen.  The function must accept arguments
+##'   \code{t}, \code{y} and \code{dydt}.  See Details for further
+##'   information.
 ##'
 ##' @param n_history Number of history points to retain.  This needs
 ##'   to be greater than zero for delay differential equations to
@@ -212,6 +225,34 @@
 ##'   vignette which shows in more detail how to solve delay
 ##'   differential equations and to use compiled objective functions.
 ##'
+##' @section Verbose output and callbacks:
+##'
+##' Debugging a failed integration can be difficult, but \code{dopri}
+##'   provides a couple of tools to get more information about where a
+##'   failure might have occured.  Most simply, one can pass
+##'   \code{verbose = TRUE} which will print information about the
+##'   time and the step size at each point just before the step is
+##'   stated.  Passing in \code{verbose = dde:::VERBOSE_EVAL} will
+##'   print information just before every evaluation of the target
+##'   function (there are several evaluations per step).
+##'
+##' However, this does not provide information about the state just
+##'   before failure.  To get that, one must provide a \code{callback}
+##'   function - this is an R function that will be called just before
+##'   a step or evaluation (based on the value of the \code{verbose}
+##'   argument) in place of the default print.  Define a callback
+##'   function with arguments \code{t}, \code{h} and \code{y} where
+##'   \code{t} is the time (beginning of a step or location of an
+##'   evaluation), \code{h} is the step size (or \code{NA} for an
+##'   evaluation) and \code{y} is the state at the point of the step
+##'   or evaluation.  Your callback function can do anything - you can
+##'   print to the screen (using \code{cat} or \code{message}), you
+##'   can store results using a closure and \code{<<-} or you could
+##'   conditionally use a \code{browser()} call to debug
+##'   interactively.  However, it is not possible for the callback to
+##'   affect the solution short of throwing an error and interrupting
+##'   it.  See the Examples for an example of use.
+##'
 ##' @examples
 ##'
 ##' # The lorenz attractor:
@@ -230,6 +271,18 @@
 ##' tt <- seq(0, 100, length.out = 40000)
 ##' y <- dde::dopri(y0, tt, lorenz, p, return_time = FALSE)
 ##' plot(y[, c(1, 3)], type = "l", lwd = 0.5, col = "#00000066")
+##'
+##' # If we want to print progress as the integration progresses we can
+##' # use the verbose argument:
+##' y <- dde::dopri(y0, c(0, 0.1), lorenz, p, verbose = TRUE)
+##'
+##' # Or print the y values too using a callback:
+##' callback <- function(t, h, y) {
+##'   message(sprintf("t: %f, h: %e, y: [%s]", t, h,
+##'                   paste(format(y, 5), collapse = ", ")))
+##' }
+##' y <- dde::dopri(y0, c(0, 0.1), lorenz, p, verbose = TRUE,
+##'                 callback = callback)
 dopri <- function(y, times, func, parms, ...,
                   n_out = 0L, output = NULL,
                   rtol = 1e-6, atol = 1e-6,
@@ -238,7 +291,7 @@ dopri <- function(y, times, func, parms, ...,
                   tcrit = NULL, event_time = NULL, event_function = NULL,
                   method = "dopri5",
                   stiff_check = 0,
-                  verbose = FALSE,
+                  verbose = FALSE, callback = NULL,
                   n_history = 0, grow_history = FALSE,
                   return_history = n_history > 0, dllname = "",
                   parms_are_real = TRUE,
@@ -285,7 +338,6 @@ dopri <- function(y, times, func, parms, ...,
   assert_size(n_history)
 
   assert_scalar_logical(grow_history)
-  assert_scalar_logical(verbose)
   assert_scalar_logical(return_history)
   assert_scalar_logical(parms_are_real)
   assert_scalar_logical(return_by_column)
@@ -295,6 +347,9 @@ dopri <- function(y, times, func, parms, ...,
   assert_scalar_logical(return_output_with_y)
   assert_scalar_logical(restartable)
   assert_size(stiff_check)
+
+  verbose <- dopri_verbose(verbose)
+  callback <- dopri_callback(callback)
 
   ynames <- check_ynames(y, ynames)
 
@@ -346,7 +401,7 @@ dopri <- function(y, times, func, parms, ...,
                ## Critical and events
                as.numeric(tcrit), is_event, event,
                ## Other:
-               use_853, as.integer(stiff_check), verbose,
+               use_853, as.integer(stiff_check), verbose, callback,
                ## Return information:
                as.integer(n_history), grow_history, return_history,
                return_initial, return_statistics, restartable)
@@ -613,8 +668,43 @@ dopri_methods <- function() {
 }
 
 
+dopri_verbose <- function(verbose) {
+  assert_scalar(verbose)
+  if (is.logical(verbose)) {
+    verbose <- if (verbose) VERBOSE_STEP else VERBOSE_QUIET
+  } else {
+    assert_integer(verbose)
+    valid <- c(VERBOSE_QUIET, VERBOSE_STEP, VERBOSE_EVAL)
+    if (!any(verbose == valid)) {
+      stop("Invalid value for verbose")
+    }
+    verbose <- as.integer(verbose)
+  }
+  verbose
+}
+
+
+dopri_callback <- function(callback) {
+  if (is.null(callback)) {
+    return(NULL)
+  }
+  if (!is.function(callback)) {
+    stop("Expected a function for 'callback'")
+  }
+  if (length(formals(callback)) != 3) {
+    stop("Expected a function with 3 arguments for 'callback'")
+  }
+  list(callback, new.env(parent = environment(callback)))
+}
+
+
 DOPRI_IDX_TARGET <- 1L
 DOPRI_IDX_PARMS <- 2L
 DOPRI_IDX_ENV <- 3L
 DOPRI_IDX_OUTPUT <- 4L
 DOPRI_IDX_EVENT <- 5L
+
+## Must match up with enum dopri_verbose in dopri.h
+VERBOSE_QUIET <- 0L
+VERBOSE_STEP <- 1L
+VERBOSE_EVAL <- 2L