+++ tests/testthat/test-ode.R
@@ -655,6 +655,98 @@ test_that("verbose mode prints trace", {
 
   expect_silent(ans1 <- dopri(y0, tt, lorenz, p))
   expect_output(ans2 <- dopri(y0, tt, lorenz, p, verbose = TRUE),
-                "t: .+, h: .+, accept: (yes|no)\n")
+                ".step. t: .+, h: .+\n")
+  expect_output(ans3 <- dopri(y0, tt, lorenz, p, verbose = VERBOSE_EVAL),
+                ".eval. t: .+, h: .+\n")
+
   expect_identical(ans1, ans2)
+  expect_identical(ans1, ans3)
+})
+
+
+test_that("verbose with callback", {
+  p <- c(10, 28, 8 / 3)
+  y0 <- c(10, 1, 1)
+
+  lorenz <- function(t, y, p) {
+    sigma <- p[[1L]]
+    R <- p[[2L]]
+    b <- p[[3L]]
+    c(sigma * (y[[2L]] - y[[1L]]),
+      R * y[[1L]] - y[[2L]] - y[[1L]] * y[[3L]],
+      -b * y[[3L]] + y[[1L]] * y[[2L]])
+  }
+
+  tt <- seq(0, 1, length.out = 200)
+  names(y0) <- letters[1:3]
+
+  steps <- list()
+  evals <- list()
+  callback <- function(t, h, y) {
+    if (is.na(h)) {
+      evals <<- c(evals, list(list(t = t, h = h, y = y)))
+    } else {
+      steps <<- c(steps, list(list(t = t, h = h, y = y)))
+    }
+  }
+
+  ## Callback successfully returns all steps - we could have done
+  ## something more interesting here, and typically this would involve
+  ## printing not collection.
+  ans1 <- dopri(y0, tt, lorenz, p, verbose = TRUE, callback = callback)
+  n <- length(steps)
+  expect_gt(n, 0)
+  expect_equal(steps[[1]]$t, 0)
+  expect_equal(steps[[1]]$y, unname(y0))
+  expect_equal(steps[[n]]$t + steps[[n]]$h, 1)
+  expect_equal(length(evals), 0)
+
+  ans2 <- dopri(y0, tt, lorenz, p, verbose = VERBOSE_EVAL, callback = callback)
+  expect_identical(steps[1:n], steps[-(1:n)])
+  m <- length(evals)
+  expect_gt(m, 0)
+  expect_equal(evals[[1]]$t, 0)
+  expect_equal(evals[[1]]$y, unname(y0))
+  expect_true(all(is.na(vapply(evals, function(x) x$h, numeric(1)))))
+
+  ## Time must have changed
+  expect_gt(evals[[2]]$t, evals[[1]]$t)
+  ## Output must have changed
+  expect_true(all(evals[[2]]$y != evals[[1]]$y))
+})
+
+
+test_that("check verbose argument", {
+  expect_identical(dopri_verbose(FALSE), VERBOSE_QUIET)
+  expect_identical(dopri_verbose(0.0), VERBOSE_QUIET)
+  expect_identical(dopri_verbose(0L), VERBOSE_QUIET)
+
+  expect_identical(dopri_verbose(TRUE), VERBOSE_STEP)
+  expect_identical(dopri_verbose(1.0), VERBOSE_STEP)
+  expect_identical(dopri_verbose(1L), VERBOSE_STEP)
+
+  expect_identical(dopri_verbose(2.0), VERBOSE_EVAL)
+  expect_identical(dopri_verbose(2L), VERBOSE_EVAL)
+
+  expect_error(dopri_verbose(-1), "Invalid value for verbose")
+  expect_error(dopri_verbose(1.5), "verbose must be integer")
+  expect_error(dopri_verbose(c(1, 2)), "verbose must be a scalar")
+  expect_error(dopri_verbose(integer(0)), "verbose must be a scalar")
+  expect_error(dopri_verbose(NULL), "verbose must be a scalar")
+})
+
+
+test_that("check callback argument", {
+  expect_null(dopri_callback(NULL))
+
+  f <- function(a, b, c) NULL
+  res <- dopri_callback(f)
+  expect_identical(f, res[[1]])
+  expect_is(res[[2]], "environment")
+  expect_equal(parent.env(res[[2]]), environment(f))
+
+  expect_error(dopri_callback(TRUE),
+               "Expected a function for 'callback'")
+  expect_error(dopri_callback(function() 1),
+               "Expected a function with 3 arguments for 'callback'")
 })