+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Initial public release
 * Validate the derivatives are finite (non-NA, non-infinite) before the first step, avoiding cryptic errors [#13](https://github.com/mrc-ide/dde/issues/13)
+* Verbose output and a callback interface to get information about each step (or each evaluation) [#12](https://github.com/mrc-ide/dde/issues/12)
 
 # 0.0.6